+++ tmp/rubycritic/bowling_game.html
@@ -47,7 +47,7 @@
     <div class="file-time">
       <span class="committed-at">
         
-          Updated <time class='js-timeago' datetime='2020-06-12 08:11:06 -0500'>2020-06-12 08:11:06 -0500</time>
+          Updated <time class='js-timeago' datetime='2020-06-17 12:45:59 -0500'>2020-06-17 12:45:59 -0500</time>
         
       </span>
     </div>
@@ -73,7 +73,7 @@
             </div>
             <div class="col-md-3">
               <div><span class="metric">N/A</span><small> complexity/method</small></div>
-              <div><span class="metric">2</span><small> churn</small></div>
+              <div><span class="metric">3</span><small> churn</small></div>
             </div>
             <div class="col-md-3">
               <div><span class="metric">0.0</span><small> complexity</small></div>
