+++ tmp/rubycritic/code_index.html
@@ -67,7 +67,7 @@
               
                 <td class="center">
                   
-                  <div class="rating b">B</div>
+                  <div class="rating a">A</div>
                 </td>
               
               <td class="left">
@@ -77,10 +77,70 @@
                   </li>
                 </ul>
               </td>
-              <td>4</td>
-              <td>67.94</td>
+              <td>5</td>
+              <td>27.12</td>
               <td>0</td>
-              <td>9</td>
+              <td>2</td>
+            </tr>
+          
+            <tr>
+              
+                <td class="center">
+                  
+                  <div class="rating a">A</div>
+                </td>
+              
+              <td class="left">
+                <ul class="nav nav-pills">
+                  <li role="presentation" >
+                    <a href="test_bowling_game.html">TestBowlingGame</a>
+                  </li>
+                </ul>
+              </td>
+              <td>4</td>
+              <td>30.2</td>
+              <td>0</td>
+              <td>1</td>
+            </tr>
+          
+            <tr>
+              
+                <td class="center">
+                  
+                  <div class="rating a">A</div>
+                </td>
+              
+              <td class="left">
+                <ul class="nav nav-pills">
+                  <li role="presentation" >
+                    <a href="rules_game.html">RulesGame</a>
+                  </li>
+                </ul>
+              </td>
+              <td>0</td>
+              <td>42.3</td>
+              <td>0</td>
+              <td>7</td>
+            </tr>
+          
+            <tr>
+              
+                <td class="center">
+                  
+                  <div class="rating b">B</div>
+                </td>
+              
+              <td class="left">
+                <ul class="nav nav-pills">
+                  <li role="presentation" >
+                    <a href="bowling_game.html">BowlingGame</a>
+                  </li>
+                </ul>
+              </td>
+              <td>3</td>
+              <td>0.0</td>
+              <td>90</td>
+              <td>1</td>
             </tr>
           
         </tbody>
