+++ tmp/rubycritic/overview.html
@@ -68,8 +68,8 @@
             
 
             <script>
-             var turbulenceData = [{"name":"Games","x":4,"y":67.94,"rating":"B"}];
-             var score = 83.02;
+             var turbulenceData = [{"name":"Games","x":5,"y":27.12,"rating":"A"},{"name":"TestBowlingGame","x":4,"y":30.2,"rating":"A"},{"name":"RulesGame","x":0,"y":42.3,"rating":"A"},{"name":"BowlingGame","x":3,"y":0.0,"rating":"B"}];
+             var score = 89.09;
             </script>
           </div>
         </div>
@@ -89,9 +89,9 @@
                     </div>
                     <div class="pull-right Summary_Content col-xs-9 col-sm-9 col-md-9 col-lg-9">
                       <ul class="list-inline">
-                        <li><h4 id="ratingAfileCount">0</h4><span>files</span></li>
-                        <li><h4 id="ratingAfileCount">0</h4><span>churns</span></li>
-                        <li><h4 id="ratingAfileCount">0</h4><span>smells</span></li>
+                        <li><h4 id="ratingAfileCount">3</h4><span>files</span></li>
+                        <li><h4 id="ratingAfileCount">9</h4><span>churns</span></li>
+                        <li><h4 id="ratingAfileCount">10</h4><span>smells</span></li>
                       </ul>
                     </div>
                   </div>
@@ -105,8 +105,8 @@
                     <div class="pull-right Summary_Content col-xs-9 col-sm-9 col-md-9 col-lg-9">
                       <ul class="list-inline">
                         <li><h4 id="ratingAfileCount">1</h4><span>files</span></li>
-                        <li><h4 id="ratingAfileCount">4</h4><span>churns</span></li>
-                        <li><h4 id="ratingAfileCount">9</h4><span>smells</span></li>
+                        <li><h4 id="ratingAfileCount">3</h4><span>churns</span></li>
+                        <li><h4 id="ratingAfileCount">1</h4><span>smells</span></li>
                       </ul>
                     </div>
                   </div>
