+++ tmp/rubycritic/simple_cov_index.html
@@ -79,6 +79,40 @@
                 <td>0%</td>
             </tr>
           
+            <tr>
+              <tr>
+                
+                  <td>
+                    <div class="rating f">F</div>
+                  </td>
+                
+                <td>
+                  <ul class="nav nav-pills">
+                    <li role="presentation" >
+                      <a href="rules_game.html">RulesGame</a>
+                    </li>
+                  </ul>
+                </td>
+                <td>0%</td>
+            </tr>
+          
+            <tr>
+              <tr>
+                
+                  <td>
+                    <div class="rating f">F</div>
+                  </td>
+                
+                <td>
+                  <ul class="nav nav-pills">
+                    <li role="presentation" >
+                      <a href="bowling_game.html">BowlingGame</a>
+                    </li>
+                  </ul>
+                </td>
+                <td>0%</td>
+            </tr>
+          
         </table>
       </div>
     </div>
