+++ tmp/rubycritic/test_bowling_game.html
@@ -47,7 +47,7 @@
     <div class="file-time">
       <span class="committed-at">
         
-          Updated <time class='js-timeago' datetime='2020-06-12 17:19:56 -0500'>2020-06-12 17:19:56 -0500</time>
+          Updated <time class='js-timeago' datetime='2020-06-17 12:45:59 -0500'>2020-06-17 12:45:59 -0500</time>
         
       </span>
     </div>
@@ -62,22 +62,22 @@
       <div class="col-md-8">
         <div class="row">
           <div class="col-md-1">
-            <div class="rating b big">
-              B
+            <div class="rating a big">
+              A
             </div>
           </div>
           <div class="code-statistics col-md-11">
             <div class="col-md-3">
-              <div><span class="metric">13</span><small> lines of codes</small></div>
+              <div><span class="metric">15</span><small> lines of codes</small></div>
               <div><span class="metric">1</span><small> methods</small></div>
             </div>
             <div class="col-md-3">
               <div><span class="metric">30.2</span><small> complexity/method</small></div>
-              <div><span class="metric">3</span><small> churn</small></div>
+              <div><span class="metric">4</span><small> churn</small></div>
             </div>
             <div class="col-md-3">
               <div><span class="metric">30.2</span><small> complexity</small></div>
-              <div><span class="metric">69</span><small> duplications</small></div>
+              <div><span class="metric">0</span><small> duplications</small></div>
             </div>
           </div>
         </div>
@@ -88,8 +88,8 @@
             <li><a href="#" id="toggle-code" class="toggle-button button">code</a></li>
             <li class="active">
               <a href="#" id="toggle-smells" class="toggle-button button">
-                4
-                smells
+                1
+                smell
               </a>
             </li>
           </ul>
@@ -99,15 +99,17 @@
 
     <code class="prettyprint linenums lang-ruby file-code js-file-code">require &#39;minitest/autorun&#39;
 require_relative &#39;games.rb&#39;
-class Test_bowling_game &lt; Minitest::Test <ol class="nocode errors smells">  <li>    <div class="description">      <div class="heading">        <span>          <i class="fa fa-warning" aria-hidden="true"></i>          <a href="https://github.com/troessner/reek/blob/master/docs/Irresponsible-Module.md" target="_blank"><b>IrresponsibleModule</b></a>        </span>      </div>      <span>Test_bowling_game has no descriptive comment</span>          </div>  </li></ol>
-  def test_process<ol class="nocode errors smells">  <li>    <div class="description">      <div class="heading">        <span>          <i class="fa fa-warning" aria-hidden="true"></i>          <a href="http://docs.seattlerb.org/flog/" target="_blank"><b>HighComplexity</b></a>        </span>      </div>      <span>Test_bowling_game#test_process has a flog score of 30</span>          </div>  </li>  <li>    <div class="description">      <div class="heading">        <span>          <i class="fa fa-warning" aria-hidden="true"></i>          <a href="https://github.com/troessner/reek/blob/master/docs/Too-Many-Statements.md" target="_blank"><b>TooManyStatements</b></a>        </span>      </div>      <span>Test_bowling_game#test_process has approx 7 statements</span>          </div>  </li></ol>
+# The test_process class is responsible for is to verify if it is taking the expected value of any request
+class TestBowlingGame &lt; Minitest::Test
+  # the test_process metthod is responsible
+  def test_process<ol class="nocode errors smells">  <li>    <div class="description">      <div class="heading">        <span>          <i class="fa fa-warning" aria-hidden="true"></i>          <a href="http://docs.seattlerb.org/flog/" target="_blank"><b>HighComplexity</b></a>        </span>      </div>      <span>TestBowlingGame#test_process has a flog score of 30</span>          </div>  </li></ol>
     games = Games.new
-    game_perfect = [10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 10, 10]<ol class="nocode errors smells">  <li>    <div class="description">      <div class="heading">        <span>          <i class="fa fa-warning" aria-hidden="true"></i>          <a href="http://docs.seattlerb.org/flay/" target="_blank"><b>DuplicateCode</b></a>        </span>      </div>      <span>Similar code found in 3 nodes</span>              <span>Locations:</span>                  <a href="test_bowling_game.html#L6" class="js-smell-location">0</a>                  <a href="test_bowling_game.html#L7" class="js-smell-location">1</a>                  <a href="test_bowling_game.html#L8" class="js-smell-location">2</a>                  </div>  </li></ol>
-    game_spare = [10, 0, 10, 0, 2, 5, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 10, 10]<ol class="nocode errors smells">  <li>    <div class="description">      <div class="heading">        <span>          <i class="fa fa-warning" aria-hidden="true"></i>          <a href="http://docs.seattlerb.org/flay/" target="_blank"><b>DuplicateCode</b></a>        </span>      </div>      <span>Similar code found in 3 nodes</span>              <span>Locations:</span>                  <a href="test_bowling_game.html#L6" class="js-smell-location">0</a>                  <a href="test_bowling_game.html#L7" class="js-smell-location">1</a>                  <a href="test_bowling_game.html#L8" class="js-smell-location">2</a>                  </div>  </li></ol>
-    game_normal = [1, 4, 4, 5, 6, 4, 5, 5, 10, 0, 0, 1, 7, 3, 6, 4, 10, 0, 2, 8, 6]<ol class="nocode errors smells">  <li>    <div class="description">      <div class="heading">        <span>          <i class="fa fa-warning" aria-hidden="true"></i>          <a href="http://docs.seattlerb.org/flay/" target="_blank"><b>DuplicateCode</b></a>        </span>      </div>      <span>Similar code found in 3 nodes</span>              <span>Locations:</span>                  <a href="test_bowling_game.html#L6" class="js-smell-location">0</a>                  <a href="test_bowling_game.html#L7" class="js-smell-location">1</a>                  <a href="test_bowling_game.html#L8" class="js-smell-location">2</a>                  </div>  </li></ol>
-    assert_equal [30, 60, 90, 120, 150, 180, 210, 240, 270, 300] ,games.balls(game_perfect)
-    assert_equal [20, 37, 44, 74, 104, 134, 164, 194, 224, 254] ,games.balls(game_spare)
-    assert_equal [5, 14, 29, 49, 60, 61, 77, 97, 117, 133] ,games.balls(game_normal)
+    game_perfect = [10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 10, 10]
+    game_spare = [10, 0, 10, 0, 2, 5, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 0, 10, 10, 10]
+    game_normal = [1, 4, 4, 5, 6, 4, 5, 5, 10, 0, 0, 1, 7, 3, 6, 4, 10, 0, 2, 8, 6]
+    assert_equal [30, 60, 90, 120, 150, 180, 210, 240, 270, 300], games.balls(game_perfect)
+    assert_equal [20, 37, 44, 74, 104, 134, 164, 194, 224, 254], games.balls(game_spare)
+    assert_equal [5, 14, 29, 49, 60, 61, 77, 97, 117, 133], games.balls(game_normal)
   end
 end
 </code>
